% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonferroni_correction.R
\name{bonferroni_correction}
\alias{bonferroni_correction}
\title{Apply Bonferroni correction to p-values}
\usage{
bonferroni_correction(data)
}
\arguments{
\item{data}{the data frame containing the test results.}
}
\value{
the data frame with adjusted p-values.
}
\description{
Apply Bonferroni correction to p-values
}
