% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_color.R
\name{fetch_color}
\alias{fetch_color}
\title{Fetch color}
\usage{
fetch_color(num, color_source = NULL)
}
\arguments{
\item{num}{number of colors to output.}

\item{color_source}{color source/palette in the dictionary (main1, main2, ect.).}
}
\value{
a vector with the HEX values for the color code.
}
\description{
Fetches colors from color dictionary.
}
