% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_shape.R
\name{fetch_shape}
\alias{fetch_shape}
\title{Fetches shapes from shape dictionary}
\usage{
fetch_shape(num)
}
\arguments{
\item{num}{number of shapes to output.}
}
\value{
a vector with the shape values.
}
\description{
Fetches shapes from shape dictionary
}
