% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepease_rRNA.R
\docType{data}
\name{genepease_rRNA}
\alias{genepease_rRNA}
\title{Example Phyloseq object from GENEPEASE II (rRNA)}
\format{
An object of class \code{phyloseq} of length 1.
}
\source{
Internal dataset
}
\usage{
genepease_rRNA
}
\description{
A `phyloseq` object containing rRNA read counts from GENEPEASE II study. The taxonomic assignment is based on the SILVA database.
https://pure.au.dk/portal/en/projects/genepease-ii-pesticide-effects-on-agricultural-soil-ecosystems
}
\keyword{datasets}
