% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_sig_results.R
\name{highlight_sig_results}
\alias{highlight_sig_results}
\title{Highlight significant results in excel}
\usage{
highlight_sig_results(wb, taxon, designs)
}
\arguments{
\item{wb}{excel workbook object.}

\item{taxon}{the taxon (e.g., phylum) to highlight results for.}

\item{designs}{a vector of test designs.}
}
\value{
modifies the excel workbook directly.
}
\description{
Highlight significant results in excel
}
