% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_anova_tukey.R
\name{perform_anova_tukey}
\alias{perform_anova_tukey}
\title{Perform ANOVA and Tukey's HSD test}
\usage{
perform_anova_tukey(data, design)
}
\arguments{
\item{data}{the data for a single taxon.}

\item{design}{the test design (e.g., "Location").}
}
\value{
a list containing ANOVA and Tukey's HSD results.
}
\description{
Perform ANOVA and Tukey's HSD test
}
