% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_permanova.R
\name{perform_permanova}
\alias{perform_permanova}
\title{Perform PERMANOVA analysis on phyloseq object}
\usage{
perform_permanova(
  physeq,
  stats_path,
  file_name = "permanova_results.xlsx",
  designs,
  signi_limit = 0.05,
  convert_to_rel = TRUE,
  dist_method = "bray",
  n_permute = 999
)
}
\arguments{
\item{physeq}{a phyloseq object.}

\item{stats_path}{a string specifying the path where the output Excel file should be saved.}

\item{file_name}{a string specifying the name of the output excel file. If not specified, the file is named permanova_results.xlsx.}

\item{designs}{a character vector specifying the model formulas (e.g., c("Concentration", "Temperature", "Concentration + Temperature")).}

\item{signi_limit}{the significance level. Default is 0.05 (5 \%).}

\item{convert_to_rel}{convert counts to relative abundances before analysis. Default is TRUE.}

\item{dist_method}{dissimilarity index used in vegdist (see vegan package). Default is "bray".}

\item{n_permute}{number of permutations. Default is 999.}
}
\value{
results are saved to the specified directory.
}
\description{
This function performs PERMANOVA (Permutational Multivariate Analysis of Variance) on a phyloseq object using multiple design formulas, and saves the results as an Excel file.
}
\examples{
#' # Data phyloseq object:
data(qaanaaq_rRNA)
phylo <- qaanaaq_rRNA

\dontrun{
perform_permanova(phylo,
                  stats_path = "results/",
                  designs = c("Direction", "Wetness"))
}
}
