% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_univariate.R
\name{perform_univariate}
\alias{perform_univariate}
\title{Perform statistical tests on phyloseq data}
\usage{
perform_univariate(
  physeq,
  stats_path,
  level_glom = "Phylum",
  designs,
  lower_limit = 0.04
)
}
\arguments{
\item{physeq}{a phyloseq object containing microbiome data.}

\item{stats_path}{a string specifying the directory to save the results.}

\item{level_glom}{name of the level to agglomerate counts. Default is "Phylum".}

\item{lower_limit}{the minimum abundance threshold for including taxa. Default is 0.04 (4%).}

\item{test_designs}{a vector of test designs (e.g., c("Concentration", "Temperature")).}
}
\value{
results are saved to the specified directory.
}
\description{
Perform statistical tests on phyloseq data
}
