% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaanaaq_annotree.R
\docType{data}
\name{qaanaaq_annotree}
\alias{qaanaaq_annotree}
\title{Example Phyloseq object from Qaanaaq (AnnoTree)}
\format{
An object of class \code{phyloseq} of length 1.
}
\source{
Internal dataset
}
\usage{
qaanaaq_annotree
}
\description{
A `phyloseq` object containing mRNA read counts from Qaanaaq study (2023) annotated via AnnoTree.
}
\keyword{datasets}
