% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaanaaq_mRNA.R
\docType{data}
\name{qaanaaq_mRNA}
\alias{qaanaaq_mRNA}
\title{Example Phyloseq object from Qaanaaq (mRNA)}
\format{
An object of class \code{phyloseq} of length 1.
}
\source{
Internal dataset
}
\usage{
qaanaaq_mRNA
}
\description{
A `phyloseq` object containing mRNA read counts from Qaanaaq study (2023) annotated using Diamond against the CAZy, NCyc and PCyc database.
}
\keyword{datasets}
