% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaanaaq_rRNA.R
\docType{data}
\name{qaanaaq_rRNA}
\alias{qaanaaq_rRNA}
\title{Example Phyloseq object from Qaanaaq (rRNA)}
\format{
An object of class \code{phyloseq} of length 1.
}
\source{
Internal dataset
}
\usage{
qaanaaq_rRNA
}
\description{
A `phyloseq` object containing rRNA read counts from Qaanaaq study (2023). The taxonomic assignment is based on the SILVA database.
}
\keyword{datasets}
