% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ENVS.R
\name{theme_ENVS}
\alias{theme_ENVS}
\title{PhyloENVS theme}
\usage{
theme_ENVS()
}
\value{
NMDS plot created with ggplot.
}
\description{
Fundamental PhyloENVS theme
}
\examples{
# Data phyloseq object:
data(qaanaaq_rRNA)
phylo <- qaanaaq_rRNA

vis_nmds(physeq = phylo,
         group_color = "Location",
         group_shape = "Transect") +
         theme_ENVS()

}
