% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ENVS_with_gridlines.R
\name{theme_ENVS_with_gridlines}
\alias{theme_ENVS_with_gridlines}
\title{PhyloENVS theme with gridlines}
\usage{
theme_ENVS_with_gridlines(both = FALSE)
}
\arguments{
\item{both}{use both vertical and horizontal gridlines. Otherwise, only horizontal gridlines are used. Default is FALSE.}
}
\value{
A custom theme (with gridlines) to be applied across package plots.
}
\description{
Fundamental PhyloENVS theme with gridlines
}
\examples{
# Data phyloseq object:
data(qaanaaq_rRNA)
phylo <- qaanaaq_rRNA

vis_nmds(physeq = phylo,
         group_color = "Location",
         group_shape = "Transect") +
         theme_ENVS_with_gridlines()

}
