% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_abundance.R
\name{vis_abundance}
\alias{vis_abundance}
\title{Plot the relative abundances for merged species of the same taxonomic rank.}
\usage{
vis_abundance(
  physeq,
  group_x,
  group_split,
  level_glom = Phylum,
  level_select = NULL,
  group_select = NULL,
  lower_limit = 2
)
}
\arguments{
\item{lower_limit}{}
}
\description{
Plot the relative abundances for merged species of the same taxonomic rank.
}
