% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nmds.R
\name{vis_nmds}
\alias{vis_nmds}
\title{Visualize NMDS}
\usage{
vis_nmds(
  physeq,
  convert_to_rel = TRUE,
  group_color,
  group_shape,
  encircle = FALSE,
  fill_circle = FALSE,
  scale_circle = 0.1,
  scale_plot = 0.1
)
}
\arguments{
\item{physeq}{a phyloseq object.}

\item{convert_to_rel}{convert counts to relative abundances. Default is TRUE.}

\item{group_color}{the parameter in metadata to color by.}

\item{group_shape}{the parameter in metadata to shape by.}

\item{encircle}{encircle the points belonging to same group (as specified by group_color). Default is FALSE.}

\item{fill_circle}{fill the encircled area. Default is FALSE.}

\item{scale_circle}{value to scale the encircle area. Default is 0.1.}

\item{scale_plot}{value to scale the plot area. Default is 0.1.}
}
\value{
NMDS plot created with ggplot.
}
\description{
Plot the data in new ordination after dimensionality reductionnmds using Non-metric MultiDimensional Scaling (NMDS) based on the Bray-Curtis distance measure.
}
\examples{
# Data phyloseq object:
data(qaanaaq_rRNA)
phylo <- qaanaaq_rRNA

vis_nmds <- function(physeq = phylo,
                     group_color = "Location",
                     group_shape = "Transect",
                     scale_circle = 0.5,
                     scale_plot = 0.5)

}
