% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_richness.R
\name{vis_richness}
\alias{vis_richness}
\title{Plot alpha diversity with Shannon and Inverse Simpson measures}
\usage{
vis_richness(physeq, group_x, group_color)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{group_x}{the x axis sample variable to plot.}

\item{group_color}{the parameter in metadata to color by.}
}
\value{
richness plots created with ggplot.
}
\description{
Plot alpha diversity with Shannon and Inverse Simpson measures
}
