% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_shepard.R
\name{vis_shepard}
\alias{vis_shepard}
\title{Plot the Shepard diagram to asses the goodness of fit for data reduction using NMDS}
\usage{
vis_shepard(physeq_rel)
}
\arguments{
\item{physeq_rel}{a phyloseq object}
}
\value{
shepard diagram created with ggplot.
}
\description{
Plot the Shepard diagram to asses the goodness of fit for data reduction using NMDS
}
