% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_phyloseq.R
\name{data_to_phyloseq}
\alias{data_to_phyloseq}
\title{Creates a phyloseq object directly from taxonomy table or gene count file}
\usage{
data_to_phyloseq(x)
}
\arguments{
\item{x}{a test varible}
}
\value{
a phyloseq object
}
\description{
Creates a phyloseq object directly from taxonomy table or gene count file
}
\examples{
data_to_phyloseq(2)
}
