% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_shape.R
\name{fetch_shape}
\alias{fetch_shape}
\title{Fetch shape}
\usage{
fetch_shape(num)
}
\arguments{
\item{num}{number of shapes to output (up to 10 is allowed).}
}
\value{
a vector with the shape values.
}
\description{
Fetches shapes from shape dictionary.
}
\examples{
fetch_shape(num = 3)

}
