% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge Data}
\usage{
merge_data(physeq, group)
}
\arguments{
\item{physeq}{a phyloseq object.}

\item{group}{variable name in the corresponding sample_data of the phyloseq object.}
}
\value{
merged phyloseq object.
}
\description{
Merge samples based on a sample variable similar to merge_samples. For numeric sample data features, values are averaged. Otherwise, the first class among unique classes are selected.
}
\examples{
# Data phyloseq object:
data(genepease_rRNA)
phylo <- genepease_rRNA

merge_data(physeq = phylo,
           group = "Group_ID")

}
