% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_permanova.R
\name{perform_permanova}
\alias{perform_permanova}
\title{Perform PERMANOVA analysis on phyloseq object}
\usage{
perform_permanova(physeq, designs, stats_path, convert_to_rel = TRUE)
}
\arguments{
\item{physeq}{A `phyloseq` object containing microbiome data.}

\item{designs}{A character vector specifying the model formulas (e.g., c("Concentration", "Temperature", "Concentration + Temperature")).}

\item{stats_path}{A string specifying the path where the output Excel file should be saved.}

\item{convert_to_rel}{convert counts to relative abundances before analysis. Default is TRUE.}
}
\value{
A data frame containing PERMANOVA results for all specified designs, with an additional column `"Model"` indicating the tested formula.
}
\description{
This function performs PERMANOVA (Permutational Multivariate Analysis of Variance) on a phyloseq object using multiple design formulas, and saves the results as an Excel file.
}
