% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_abundance.R
\name{vis_abundance}
\alias{vis_abundance}
\title{Visualize abundance}
\usage{
vis_abundance(
  physeq,
  group_x,
  group_split,
  level_glom = Phylum,
  level_select = NULL,
  group_select = NULL,
  lower_limit = 2,
  color_source = "AU2",
  remove_grid = FALSE
)
}
\arguments{
\item{physeq}{a phyloseq object.}

\item{group_x}{the x axis sample variable to plot.}

\item{group_split}{name of the sample data column to group by.}

\item{level_glom}{name of the level to agglomerate counts. Default is "Phylum".}

\item{level_select}{specified level to target for specified group (e.g., domain level - useful with add_level()). Default is NULL.}

\item{group_select}{specified group to target in specified level (e.g., Prokaryotes in the domain level). Default is NULL.}

\item{lower_limit}{relative abundance threshold in percentages for visualizing the abundance. Abundances less than the threshold is pooled together.}

\item{remove_grid}{remove the grid splitting groups. Default is FALSE.}
}
\value{
a abundance plot created with ggplot.
}
\description{
Plot the relative abundances for merged species of the same taxonomic rank.
}
