% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nmds.R
\name{vis_nmds}
\alias{vis_nmds}
\title{Plot the data i the new ordination after data reduction using NMDS}
\usage{
vis_nmds(
  physeq,
  convert_to_rel = TRUE,
  group_color,
  group_shape,
  encircle = FALSE,
  fill_circle = FALSE,
  scale_circle = 0.1
)
}
\arguments{
\item{physeq}{a phyloseq object.}

\item{convert_to_rel}{convert counts to relative abundances. Default is TRUE.}

\item{group_color}{the parameter in metadata to color by.}

\item{group_shape}{the parameter in metadata to shape by.}

\item{encircle}{encircle the points belonging to same group (as specified by group_color). Default is FALSE.}

\item{fill_circle}{fill the encircled area. Default is FALSE.}

\item{scale_circle}{value to scale the encircle area. Default is 0.1.}
}
\value{
NMDS plot created with ggplot.
}
\description{
Plot the data i the new ordination after data reduction using NMDS
}
