% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nmds.R
\name{vis_nmds}
\alias{vis_nmds}
\title{Plot the data i the new ordination after data reduction using NMDS}
\usage{
vis_nmds(physeq_rel, group_color, group_shape)
}
\arguments{
\item{physeq_rel}{a phyloseq object}

\item{group_color}{the parameter in metadata to color by.}

\item{group_shape}{the parameter in metadata to shape by.}
}
\value{
NMDS plot created with ggplot.
}
\description{
Plot the data i the new ordination after data reduction using NMDS
}
\examples{
vis_nmds(physeq_rel, group_color, group_shape)
}
