% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_richness.R
\name{vis_richness}
\alias{vis_richness}
\title{Visualize richness plot}
\usage{
vis_richness(
  physeq,
  group_x,
  group_color,
  measures = c("Shannon", "InvSimpson")
)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{group_x}{the x axis sample variable to plot.}

\item{group_color}{the parameter in metadata to color by.}

\item{measures}{character vector of diversity indices names. Default is c("Simpson", "InvSimpson"). Supported: c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher").}
}
\value{
richness plots created with ggplot.
}
\description{
Plot alpha diversity with Shannon and Inverse Simpson measures.
}
\examples{
# Data phyloseq object:
data(qaanaaq_rRNA)
phylo <- qaanaaq_rRNA

vis_richness(physeq = phylo,
             group_x = "Location",
             group_color = "Transect")

}
