% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_richness.R
\name{vis_richness}
\alias{vis_richness}
\title{Plot alpha diversity with Shannon and Inverse Simpson measures}
\usage{
vis_richness(physeq, group_x, group_color)
}
\arguments{
\item{group_color}{}
}
\description{
Plot alpha diversity with Shannon and Inverse Simpson measures
}
