% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_shepard.R
\name{vis_shepard}
\alias{vis_shepard}
\title{Visualize Shepard diagram}
\usage{
vis_shepard(physeq, convert_to_rel = TRUE)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{convert_to_rel}{convert counts to relative abundances. Default is TRUE.}
}
\value{
shepard diagram created with ggplot.
}
\description{
Plot the Shepard diagram to asses the goodness of fit for data reduction using NMDS.
}
